backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==8.0.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
lxml==5.3.2
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
packaging==24.2
pillow==11.2.1
primp==0.14.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
smolagents @ git+https://github.com/huggingface/smolagents.git@4f877034dc7c24377c380edc1d344d3086a5c60c
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.1.0 # Added SSE library
starlette==0.46.1
tinydb==4.8.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.19.0
